fix(runtime/int64): compilation issue on Linux

+++ src/runtime/int64.h
@@ -4,9 +4,10 @@ Released under Apache 2.0 license as described in the file LICENSE.
 
 Author: Leonardo de Moura
 */
+#include <stdint.h>
 namespace lean {
-typedef long long int64;
-typedef unsigned long long uint64;
-static_assert(sizeof(int64 ) == 8,  "unexpected int64 size");
-static_assert(sizeof(uint64 ) == 8, "unexpected uint64 size");
+typedef int64_t  int64;
+typedef uint64_t uint64;
+static_assert(sizeof(int64) == 8,  "unexpected int64 size");
+static_assert(sizeof(uint64) == 8, "unexpected uint64 size");
 }
